Add files via upload
Financial improvement of each country is generally reliant on its currencies and each individual is the part of Economy. At present, in cycle of demonetization, there are long queues waiting in the banks and ATM Machines but at the other hand, many corrupted individuals are giving the cash straightforwardly by underhanded sources and it is straightforwardly affecting on monetary status of India. Numerous scientists have been worked hard to create hearty and effective automated currency detection machines. Programmed machine which can recognize banknotes are currently generally utilized in distributors of present-day items like candies, cool drinks, ticket bookings etc. The innovation of money acknowledgment fundamentally targets recognizing and extracting visible and invisible highlights of currencies. As of recently, numerous methods have been proposed to distinguish the cash note. In any case, this way isn't useful if the note is filthy or torn. In the event that a note is dirty, its shading characteristics are changed generally. So, it is significant those how we extract the highlights from the picture of currency notes and apply algorithms to improve exactness to perceive the note. The fake currency detection implementation was made in MATLAB using image processing techniques. The proposed system from image acquisition to the intensity calculation of the extracted feature. It removes the major problem by extracting features even if the currency is torn or scribbled. The algorithm proposed is also suitable for new 500 and 2000 rupees denominations. The proposed system can be implemented with hardware using suitable processors which will result in the increase of processing speed.